+++ src/command.rs
@@ -16,6 +16,9 @@ pub enum Message {
     Run(String),
     Stop,
     Finished,
+    AlreadyExists,
+    PlaylistNotChecked,
+    Error(String),
 }
 
 #[derive(Default)]
@@ -39,9 +42,9 @@ impl Command {
 
     pub fn start(
         &mut self,
-        args: Vec<String>,
+        mut args: Vec<&str>,
         show_modal: &mut bool,
-        ui_message: &mut String,
+        modal_body: &mut String,
         bin_dir: Option<PathBuf>,
         sender: Option<UnboundedSender<String>>,
     ) {
@@ -53,6 +56,28 @@ impl Command {
             command.creation_flags(CREATE_NO_WINDOW);
         }
 
+        let print = [
+            "--print",
+            r#"before_dl:__{"type": "pre_download", "video_id": "%(id)s"}"#,
+            "--print",
+            r#"playlist:__{"type": "end_of_playlist"}"#,
+            "--print",
+            r#"after_video:__{"type": "end_of_video"}"#,
+        ];
+
+        let progess_template = [
+            "--progress-template",
+            // format progress as a simple json
+            r#"__{"type": "downloading", "video_title": "%(info.title)s", "eta": %(progress.eta)s, "downloaded_bytes": %(progress.downloaded_bytes)s, "total_bytes": %(progress.total_bytes)s, "elapsed": %(progress.elapsed)s, "speed": %(progress.speed)s, "playlist_count": %(info.playlist_count)s, "playlist_index": %(info.playlist_index)s }"#,
+            // "--progress-template",
+            // r#"postprocess:__{"type": "post_processing", "status": "%(progress.status)s"}"#,
+        ];
+
+        args.extend_from_slice(&print);
+        args.extend_from_slice(&progess_template);
+
+        args.push("--no-quiet");
+
         let Ok(shared_child) = SharedChild::spawn(
             command
                 .args(args)
@@ -61,7 +86,7 @@ impl Command {
         ) else {
             tracing::error!("Spawning child process failed");
             *show_modal = true;
-            *ui_message = String::from("yt-dlp binary is missing");
+            *modal_body = String::from("yt-dlp binary is missing");
             return;
         };
 
@@ -69,26 +94,26 @@ impl Command {
 
         let Some(child) = self.shared_child.clone() else {
                         *show_modal = true;
-                        *ui_message = String::from("Something went wrong");
+                        *modal_body = String::from("Something went wrong");
                         tracing::error!("No child process");
                         return;
                     };
 
         *show_modal = true;
-        *ui_message = String::from("Initializing...");
+        *modal_body = String::from("Initializing...");
 
         if let Some(stderr) = child.take_stderr() {
             let Some(sender) = sender.clone() else {
                 *show_modal = true;
-                *ui_message = String::from("Something went wrong");
+                *modal_body = String::from("Something went wrong");
                 return;
             };
             std::thread::spawn(move || {
                 let reader = BufReader::new(stderr);
                 for line in reader.lines().flatten() {
                     sender
-                        .unbounded_send(line)
-                        .unwrap_or_else(|e| tracing::error!("{e}"));
+                        .unbounded_send(format!("stderr:{line}"))
+                        .unwrap_or_else(|e| tracing::error!("failed to send stderr: {e}"));
                 }
             });
         }
@@ -96,12 +121,12 @@ impl Command {
         if let Some(stdout) = child.take_stdout() {
             let Some(sender) = sender else {
                 *show_modal = true;
-                *ui_message = String::from("Something went wrong");
+                *modal_body = String::from("Something went wrong");
                 return;
             };
             std::thread::spawn(move || {
                 let mut reader = BufReader::new(stdout);
-                let mut buffer: Vec<u8> = Vec::new();
+                let mut buffer = vec![];
                 loop {
                     let Ok(bytes_read) = reader.read_until(b'\r', &mut buffer) else {
                                         panic!("failed to read buffer");
@@ -111,21 +136,15 @@ impl Command {
                         break;
                     }
 
-                    match std::str::from_utf8(&buffer) {
-                        Ok(str) => {
-                            sender
-                                .unbounded_send(str.to_string())
-                                .unwrap_or_else(|e| tracing::debug!("{e}"));
-                        }
-                        Err(err) => {
-                            tracing::debug!("{err}");
-                        }
-                    }
+                    sender
+                        .unbounded_send(String::from_utf8_lossy(&buffer).to_string())
+                        .unwrap_or_else(|e| tracing::error!("failed to send stdout: {e}"));
+
                     buffer.clear();
                 }
-                sender
-                    .unbounded_send(String::from("Finished"))
-                    .unwrap_or_else(|_e| tracing::error!("{_e}"));
+                // sender
+                //     .unbounded_send(String::from("Finished"))
+                //     .unwrap_or_else(|e| tracing::error!(r#"failed to send "Finished": {e}"#));
             });
         }
     }